8.18.1

+++ src/main/java/com/geeksun/jvm/instructions/stores/IStore0.java
@@ -13,6 +13,6 @@ public class IStore0 extends NoOperandsInstruction {
     @Override
     public void execute(Frame frame) {
         int val = frame.getOperandStack().popInt();
-        frame.getLocalVars().setLong(0, val);
+        frame.getLocalVars().setInt(0, val);
     }
 }
